\name{CN}
\docType{data}
\alias{CN}
\title{A CNAnorm object with information about most abbundant ploidy states, obtained from data LS041 }
\description{
This data is to provide an object to use in several example without having to wait for computing it.
To see how it was generated, see documentation of function peakPloidy 
}
\usage{data(CN)}
\format{A CNAnorm object}
% \references{
% McNeil, D. R. (1977) \emph{Interactive Data Analysis}.
% New York: Wiley.
% }
\keyword{datasets, object}
