\name{hg19_hs_ideogr}
\docType{data}
\alias{hg19_hs_ideogr}
\title{An object with the ideogram information for homo sapiens - hg19}

\description{
This is bundles data that can be provided to plotGenome in order to plot
location of the centromere. In future release it might be used to produce
an ideogram too}



\usage{data(hg19_hs_ideogr)}
\format{A data.frame}
% \references{
% McNeil, D. R. (1977) \emph{Interactive Data Analysis}.
% New York: Wiley.
% }
\keyword{datasets, object}
