\name{LS041}
\docType{data}
\alias{LS041}
\title{Mapped reads in tumor and matched blood for patient LS041 }
\description{
This data set provide reads in tumor and matched blood for patient LS041.
Each row has information about non-overlapping window across the genome. In particular
it reports: chromosome name, starting position of the window (1 based), 
number of mapped reads in the test (lung tumor), number of reads in the control (matched blood)
and GC content of the window.
}
\usage{data(LS041)}
\format{A dataframe}

\references{ Gusnanto, A., Wood, H.M., Pawitan, Y., Rabbitts, P. and Berri, S.
(2011) \emph{Correcting for cancer genome size and tumour cell content enables
better estimation of copy number alterations from next generation sequence
data.} Bioinformatics}


\keyword{datasets, dataframe}
