\name{LS041}
\docType{data}
\alias{LS041}
\title{Mapped reads in tumour and matched blood for patient LS041 }
\description{
This data set provide reads in tumour and mathced blood for patient LS041.
Each row has information about non-overlapping window across the genome. In particular
it reports: chromosome name, starting position of the window (1 based), 
number of mapped reads in the test (lung tumour), number of reads in the control (matched blood)
and GC content of the window.
}
\usage{data(LS041)}
\format{A dataframe}
% \references{
% McNeil, D. R. (1977) \emph{Interactive Data Analysis}.
% New York: Wiley.
% }
\keyword{datasets, dataframe}
